+++ apps/frontend/src/app/account/de.json
@@ -1,0 +1,7 @@
+{
+  "title": "Konto",
+  "sections": {
+    "profile": "Profil",
+    "security": "Sicherheit"
+  }
+}
