+++ apps/frontend/src/app/account/en.json
@@ -1,0 +1,7 @@
+{
+  "title": "Account",
+  "sections": {
+    "profile": "Profile",
+    "security": "Security"
+  }
+}
