+++ apps/frontend/src/app/account/username/de.json
@@ -5,5 +5,11 @@
   "errors": {
     "updateFailed": "Aktualisierung des Benutzernamens fehlgeschlagen",
     "oncePerDay": "Der Benutzername kann nur einmal pro Tag geändert werden"
+  },
+  "username": {
+    "label": "Benutzername"
+  },
+  "actions": {
+    "save": "Speichern"
   }
 }
